+++ src/Page/GroupPage/Post/PostBarrel.ts
@@ -1,0 +1,11 @@
+import axios from 'axios';
+
+import { styled } from 'styled-components';
+import { useState } from 'react';
+import PostHeader from '../PostHeader';
+import PostContent from './PostContent';
+import { PostProps } from '../PostPropsType';
+
+export { styled, useState };
+export { PostHeader, PostContent };
+export type {PostProps}
