+++ src/Page/GroupPage/Post/PostContent.tsx
@@ -1,10 +1,10 @@
 import { styled } from 'styled-components';
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { putEditBoard } from '../../API/Boards';
+import { putEditBoard } from '../../../API/Boards';
 import { useParams } from 'react-router-dom';
-import { PostProps } from './PostPropsType';
+import { PostProps } from '../PostPropsType';
 import { useRecoilState } from 'recoil';
-import { refreshState } from '../../data/initialData';
+import { refreshState } from '../../../data/initialData';
 interface PostContentProps {
   editbind: { edit: boolean; setEdit: React.Dispatch<React.SetStateAction<boolean>> };
   postProps: PostProps;
